@@ -1,28 +1,34 @@
 name: firebase_admob
 description: Flutter plugin for Firebase AdMob, supporting
   banner, interstitial (full-screen), and rewarded video ads
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_admob
-version: 0.9.0+9
+version: 0.9.2
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.firebaseadmob
-    iosPrefix: FLT
-    pluginClass: FirebaseAdMobPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.firebaseadmob
+        pluginClass: FirebaseAdMobPlugin
+      ios:
+        pluginClass: FLTFirebaseAdMobPlugin
 
 dependencies:
   meta: ^1.0.4
   platform: ^2.0.0
   flutter:
     sdk: flutter
+  firebase_core: ^0.4.2+1
 
 dev_dependencies:
+  pedantic: ^1.8.0
+  e2e: ^0.2.1
+  flutter_driver:
+    sdk: flutter
   mockito: ^3.0.0
   flutter_test:
     sdk: flutter
-  firebase_core: ^0.4.0
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.5.0 <2.0.0"
+  flutter: ">=1.10.0 <2.0.0"
