@@ -1,15 +1,17 @@
 name: firebase_analytics
 description: Flutter plugin for Google Analytics for Firebase, an app measurement
   solution that provides insight on app usage and user engagement on Android and iOS.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_analytics/firebase_analytics
-version: 5.0.8
+version: 5.0.12
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.firebaseanalytics
-    iosPrefix: FLT
-    pluginClass: FirebaseAnalyticsPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.firebaseanalytics
+        pluginClass: FirebaseAnalyticsPlugin
+      ios:
+        pluginClass: FLTFirebaseAnalyticsPlugin
 
 dependencies:
   meta: ^1.0.4
@@ -17,6 +19,7 @@ dependencies:
     sdk: flutter
 
 dev_dependencies:
+  pedantic: ^1.8.0
   e2e: ^0.2.1
   mockito: 3.0.0
   flutter_test:
@@ -27,4 +30,4 @@ dev_dependencies:
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.4 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
