@@ -1,8 +1,7 @@
 name: firebase_auth_web
 description: The web implementation of firebase_auth
-author: Flutter Team <flutter-dev@googlegroups.com>
-homepage: https://github.com/FirebaseExtended/flutterfire/packages/firebase_auth/firebase_auth_web
-version: 0.1.0
+homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_auth/firebase_auth_web
+version: 0.1.2+2
 
 flutter:
   plugin:
@@ -18,19 +17,20 @@ dependencies:
   flutter_web_plugins:
     sdk: flutter
   firebase: ^7.0.0
+  http_parser: ^3.1.3
   meta: ^1.1.7
   js: ^0.6.1
 
 dev_dependencies:
+  pedantic: ^1.8.0
   flutter_test:
     sdk: flutter
   firebase_auth:
     path: ../firebase_auth
   firebase_core_platform_interface: ^1.0.0
-  firebase_core_web:
-    path: ../../firebase_core/firebase_core_web
+  firebase_core_web: ^0.1.1+2
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.2 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
 
