@@ -1,30 +1,44 @@
 name: firebase_core
 description: Flutter plugin for Firebase Core, enabling connecting to multiple
   Firebase apps.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_core/firebase_core
-version: 0.4.2+1
+version: 0.4.4+2
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.firebase.core
-    iosPrefix: FLT
-    pluginClass: FirebaseCorePlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.firebase.core
+        pluginClass: FirebaseCorePlugin
+      ios:
+        pluginClass: FLTFirebaseCorePlugin
+      macos:
+        pluginClass: FLTFirebaseCorePlugin
+      web:
+        default_package: firebase_core_web
 
 dependencies:
-  firebase_core_platform_interface: ^1.0.0
+  firebase_core_platform_interface: ^1.0.4
   flutter:
     sdk: flutter
   meta: "^1.0.5"
+  # The design on https://flutter.dev/go/federated-plugins was to leave
+  # this constraint as "any". We cannot do it right now as it fails pub publish
+  # validation, so we set a ^ constraint.
+  # TODO(amirh): Revisit this (either update this part in the  design or the pub tool).
+  # https://github.com/flutter/flutter/issues/46264
+  firebase_core_web: ^0.1.1
 
 dev_dependencies:
+  pedantic: ^1.8.0
   e2e: ^0.2.1
   flutter_driver:
     sdk: flutter
   flutter_test:
     sdk: flutter
   mockito: ^4.1.1
+  plugin_platform_interface: ^1.0.2
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.5 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
