@@ -1,8 +1,7 @@
 name: firebase_core_web
 description: The web implementation of firebase_core
-author: Flutter Team <flutter-dev@googlegroups.com>
-homepage: https://github.com/FirebaseExtended/flutterfire/packages/firebase_core/firebase_core_web
-version: 0.1.0+2
+homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_core/firebase_core_web
+version: 0.1.1+3
 
 flutter:
   plugin:
@@ -22,10 +21,11 @@ dependencies:
   js: ^0.6.1
 
 dev_dependencies:
+  pedantic: ^1.8.0
   flutter_test:
     sdk: flutter
   firebase_core: ^0.4.2+1
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.2 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
