@@ -2,13 +2,12 @@ name: firebase_crashlytics
 description:
   Flutter plugin for Firebase Crashlytics. It reports uncaught errors to the
   Firebase console.
-version: 0.1.2
-author: Flutter Team <flutter-dev@google.com>
+version: 0.1.3+2
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_crashlytics
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.5 <2.0.0"
+  flutter: ">=1.12.13+hotfix.4 <2.0.0"
 
 dependencies:
   flutter:
@@ -16,6 +15,7 @@ dependencies:
   stack_trace: ^1.9.3
 
 dev_dependencies:
+  pedantic: ^1.8.0
   e2e: ^0.2.1
   flutter_test:
     sdk: flutter
@@ -25,5 +25,11 @@ dev_dependencies:
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.firebase.crashlytics.firebasecrashlytics
-    pluginClass: FirebaseCrashlyticsPlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.firebase.crashlytics.firebasecrashlytics
+        pluginClass: FirebaseCrashlyticsPlugin
+      ios:
+        pluginClass: FirebaseCrashlyticsPlugin
+      macos:
+        pluginClass: FirebaseCrashlyticsPlugin
