@@ -1,15 +1,18 @@
 name: firebase_database
 description: Flutter plugin for Firebase Database, a cloud-hosted NoSQL database
   with realtime data syncing across Android and iOS clients, and offline access.
-author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/FirebaseExtended/flutterfire/tree/master/packages/firebase_database
-version: 3.1.0
+version: 3.1.3
 
 flutter:
   plugin:
-    androidPackage: io.flutter.plugins.firebase.database
-    iosPrefix: FLT
-    pluginClass: FirebaseDatabasePlugin
+    platforms:
+      android:
+        package: io.flutter.plugins.firebase.database
+        pluginClass: FirebaseDatabasePlugin
+      ios:
+        pluginClass: FLTFirebaseDatabasePlugin
+
 
 dependencies:
   flutter:
@@ -17,6 +20,7 @@ dependencies:
   firebase_core: ^0.4.0
 
 dev_dependencies:
+  pedantic: ^1.8.0
   test: ^1.3.0
   mockito: ^3.0.0
   flutter_test:
@@ -27,4 +31,4 @@ dev_dependencies:
 
 environment:
   sdk: ">=2.0.0-dev.28.0 <3.0.0"
-  flutter: ">=1.9.1+hotfix.2 <2.0.0"
+  flutter: ">=1.10.0 <2.0.0"
