@@ -187,11 +187,11 @@ class AdSize {
 /// A valid [adUnitId] is required.
 abstract class MobileAd {
   /// Default constructor, used by subclasses.
-  MobileAd(
-      {@required this.adUnitId,
-      MobileAdTargetingInfo targetingInfo,
-      this.listener})
-      : _targetingInfo = targetingInfo ?? const MobileAdTargetingInfo() {
+  MobileAd({
+    @required this.adUnitId,
+    MobileAdTargetingInfo targetingInfo,
+    this.listener,
+  }) : _targetingInfo = targetingInfo ?? const MobileAdTargetingInfo() {
     assert(adUnitId != null && adUnitId.isNotEmpty);
     assert(_allAds[id] == null);
     _allAds[id] = this;
@@ -275,8 +275,12 @@ class BannerAd extends MobileAd {
 
   final AdSize size;
 
-  /// These are AdMob's test ad unit IDs, which always return test ads. You're
-  /// encouraged to use them for testing in your own apps.
+  /// {@template firebase_admob.testAdUnitId}
+  /// A platform-specific AdMob test ad unit ID. This ad unit
+  /// has been specially configured to always return test ads, and developers
+  /// are encouraged to use it while building and testing their apps.
+  /// {@endtemplate}
+  /// {@macro firebase_admob.testAdUnitId}
   static final String testAdUnitId = Platform.isAndroid
       ? 'ca-app-pub-3940256099942544/6300978111'
       : 'ca-app-pub-3940256099942544/2934735716';
@@ -290,6 +294,68 @@ class BannerAd extends MobileAd {
       'width': size.width,
       'height': size.height,
       'adSizeType': size.adSizeType.toString(),
+    });
+  }
+}
+
+/// A NativeAd for the [FirebaseAdMobPlugin].
+///
+/// Native ads are ad assets that are presented to users via UI components that
+/// are native to the platform. (e.g. A
+/// [View](https://developer.android.com/reference/android/view/View) on Android
+/// or a
+/// [UIView](https://developer.apple.com/documentation/uikit/uiview?language=objc)
+/// on iOS). Using Flutter widgets to create native ads is NOT supported by
+/// this.
+///
+/// Using platform specific UI components, these ads can be formatted to match
+/// the visual design of the user experience in which they live. In coding
+/// terms, this means that when a native ad loads, your app receives a NativeAd
+/// object that contains its assets, and the app (rather than the Google Mobile
+/// Ads SDK) is then responsible for displaying them.
+///
+/// See the README for more details on using Native Ads.
+///
+/// Note: This is currently only supported for Android.
+class NativeAd extends MobileAd {
+  NativeAd({
+    @required String adUnitId,
+    @required this.factoryId,
+    MobileAdTargetingInfo targetingInfo,
+    MobileAdListener listener,
+    this.customOptions,
+  }) : super(
+            adUnitId: adUnitId,
+            targetingInfo: targetingInfo,
+            listener: listener) {
+    if (!Platform.isAndroid) {
+      throw UnimplementedError(
+        '$NativeAd is currently only available for Android',
+      );
+    }
+  }
+
+  /// Optional options used to create the [NativeAd].
+  ///
+  /// These options are passed to the platform's `NativeAdFactory`.
+  final Map<String, dynamic> customOptions;
+
+  /// An identifier for the factory that creates the Platform view.
+  final String factoryId;
+
+  /// {@macro firebase_admob.testAdUnitId}
+  static final String testAdUnitId = Platform.isAndroid
+      ? 'ca-app-pub-3940256099942544/2247696110'
+      : 'ca-app-pub-3940256099942544/3986624511';
+
+  @override
+  Future<bool> load() {
+    return _invokeBooleanMethod("loadNativeAd", <String, dynamic>{
+      'id': id,
+      'factoryId': factoryId,
+      'targetingInfo': targetingInfo?.toJson(),
+      'adUnitId': adUnitId,
+      'customOptions': customOptions,
     });
   }
 }
@@ -308,9 +374,7 @@ class InterstitialAd extends MobileAd {
             targetingInfo: targetingInfo,
             listener: listener);
 
-  /// A platform-specific AdMob test ad unit ID for interstitials. This ad unit
-  /// has been specially configured to always return test ads, and developers
-  /// are encouraged to use it while building and testing their apps.
+  /// {@macro firebase_admob.testAdUnitId}
   static final String testAdUnitId = Platform.isAndroid
       ? 'ca-app-pub-3940256099942544/1033173712'
       : 'ca-app-pub-3940256099942544/4411468910';
@@ -400,6 +464,31 @@ class RewardedVideoAd {
   /// Callback invoked for events in the rewarded video ad lifecycle.
   RewardedVideoAdListener listener;
 
+  String _userId;
+  String _customData;
+
+  /// The user id used in server-to-server reward callbacks
+  String get userId => _userId;
+
+  /// The custom data included in server-to-server reward callbacks
+  String get customData => _customData;
+
+  /// Sets the user id to be used in server-to-server reward callbacks.
+  set userId(String userId) {
+    _invokeBooleanMethod("setRewardedVideoAdUserId", <String, dynamic>{
+      'userId': userId,
+    });
+    _userId = userId;
+  }
+
+  /// Sets custom data to be included in server-to-server reward callbacks.
+  set customData(String customData) {
+    _invokeBooleanMethod("setRewardedVideoAdCustomData", <String, dynamic>{
+      'customData': customData,
+    });
+    _customData = customData;
+  }
+
   /// Shows a rewarded video ad if one has been loaded.
   Future<bool> show() {
     return _invokeBooleanMethod("showRewardedVideoAd");
@@ -444,9 +533,7 @@ class FirebaseAdMob {
     _channel.setMethodCallHandler(_handleMethod);
   }
 
-  // A placeholder AdMob App ID for testing. AdMob App IDs and ad unit IDs are
-  // specific to a single operating system, so apps building for both Android and
-  // iOS will need a set for each platform.
+  /// {@macro firebase_admob.testAdUnitId}
   static final String testAppId = Platform.isAndroid
       ? 'ca-app-pub-3940256099942544~3347511713'
       : 'ca-app-pub-3940256099942544~1458002511';
