@@ -9,6 +9,7 @@ import 'package:firebase_core/firebase_core.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
 
 Future<void> main() async {
+  WidgetsFlutterBinding.ensureInitialized();
   final FirebaseApp app = await FirebaseApp.configure(
     name: 'test',
     options: const FirebaseOptions(
@@ -32,7 +33,10 @@ class MessageList extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return StreamBuilder<QuerySnapshot>(
-      stream: firestore.collection('messages').snapshots(),
+      stream: firestore
+          .collection("messages")
+          .orderBy("created_at", descending: true)
+          .snapshots(),
       builder: (BuildContext context, AsyncSnapshot<QuerySnapshot> snapshot) {
         if (!snapshot.hasData) return const Text('Loading...');
         final int messageCount = snapshot.data.documents.length;
@@ -42,6 +46,10 @@ class MessageList extends StatelessWidget {
             final DocumentSnapshot document = snapshot.data.documents[index];
             final dynamic message = document['message'];
             return ListTile(
+              trailing: IconButton(
+                onPressed: () => document.reference.delete(),
+                icon: Icon(Icons.delete),
+              ),
               title: Text(
                 message != null ? message.toString() : '<No message retrieved>',
               ),
@@ -68,11 +76,77 @@ class MyHomePage extends StatelessWidget {
     });
   }
 
+  Future<void> _runTransaction() async {
+    firestore.runTransaction((Transaction transaction) async {
+      final allDocs = await firestore.collection("messages").getDocuments();
+      final toBeRetrieved =
+          allDocs.documents.sublist(allDocs.documents.length ~/ 2);
+      final toBeDeleted =
+          allDocs.documents.sublist(0, allDocs.documents.length ~/ 2);
+      await Future.forEach(toBeDeleted, (DocumentSnapshot snapshot) async {
+        await transaction.delete(snapshot.reference);
+      });
+
+      await Future.forEach(toBeRetrieved, (DocumentSnapshot snapshot) async {
+        await transaction.update(snapshot.reference, {
+          "message": "Updated from Transaction",
+          "created_at": FieldValue.serverTimestamp()
+        });
+      });
+    });
+
+    await Future.forEach(List.generate(2, (index) => index), (item) async {
+      await firestore.runTransaction((Transaction transaction) async {
+        await Future.forEach(List.generate(10, (index) => index), (item) async {
+          await transaction.set(firestore.collection("messages").document(), {
+            "message": "Created from Transaction $item",
+            "created_at": FieldValue.serverTimestamp()
+          });
+        });
+      });
+    });
+  }
+
+  Future<void> _runBatchWrite() async {
+    final batchWrite = firestore.batch();
+    final querySnapshot = await firestore
+        .collection("messages")
+        .orderBy("created_at")
+        .limit(12)
+        .getDocuments();
+    querySnapshot.documents
+        .sublist(0, querySnapshot.documents.length - 3)
+        .forEach((DocumentSnapshot doc) {
+      batchWrite.updateData(doc.reference, {
+        "message": "Batched message",
+        "created_at": FieldValue.serverTimestamp()
+      });
+    });
+    batchWrite.setData(firestore.collection("messages").document(), {
+      "message": "Batched message created",
+      "created_at": FieldValue.serverTimestamp()
+    });
+    batchWrite.delete(
+        querySnapshot.documents[querySnapshot.documents.length - 2].reference);
+    batchWrite.delete(querySnapshot.documents.last.reference);
+    await batchWrite.commit();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
         title: const Text('Firestore Example'),
+        actions: <Widget>[
+          FlatButton(
+            onPressed: _runTransaction,
+            child: Text("Run Transaction"),
+          ),
+          FlatButton(
+            onPressed: _runBatchWrite,
+            child: Text("Batch Write"),
+          )
+        ],
       ),
       body: MessageList(firestore: firestore),
       floatingActionButton: FloatingActionButton(
