@@ -5,32 +5,25 @@
 library cloud_firestore;
 
 import 'dart:async';
-import 'dart:convert';
-import 'dart:typed_data';
-import 'dart:ui' show hashValues, hashList;
+import 'dart:ui' show hashList;
 
-import 'package:collection/collection.dart';
+import 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart'
+    as platform;
 import 'package:firebase_core/firebase_core.dart';
-import 'package:flutter/foundation.dart' show ReadBuffer, WriteBuffer;
-import 'package:flutter/services.dart';
-import 'package:meta/meta.dart';
 
-import 'src/utils/auto_id_generator.dart';
+export 'package:cloud_firestore_platform_interface/cloud_firestore_platform_interface.dart'
+    show FieldPath, Blob, GeoPoint, Timestamp, Source, DocumentChangeType;
 
-part 'src/blob.dart';
 part 'src/collection_reference.dart';
 part 'src/document_change.dart';
+part 'src/utils/platform_utils.dart';
 part 'src/document_reference.dart';
 part 'src/document_snapshot.dart';
-part 'src/field_path.dart';
 part 'src/field_value.dart';
 part 'src/firestore.dart';
-part 'src/firestore_message_codec.dart';
-part 'src/geo_point.dart';
 part 'src/query.dart';
 part 'src/query_snapshot.dart';
+part 'src/utils/codec_utility.dart';
 part 'src/snapshot_metadata.dart';
-part 'src/timestamp.dart';
 part 'src/transaction.dart';
 part 'src/write_batch.dart';
-part 'src/source.dart';
