@@ -4,56 +4,36 @@
 
 part of cloud_firestore;
 
-/// An enumeration of document change types.
-enum DocumentChangeType {
-  /// Indicates a new document was added to the set of documents matching the
-  /// query.
-  added,
-
-  /// Indicates a document within the query was modified.
-  modified,
-
-  /// Indicates a document within the query was removed (either deleted or no
-  /// longer matches the query.
-  removed,
-}
-
 /// A DocumentChange represents a change to the documents matching a query.
 ///
 /// It contains the document affected and the type of change that occurred
 /// (added, modified, or removed).
 class DocumentChange {
-  DocumentChange._(Map<dynamic, dynamic> data, Firestore firestore)
-      : oldIndex = data['oldIndex'],
-        newIndex = data['newIndex'],
-        document = DocumentSnapshot._(
-          data['path'],
-          _asStringKeyedMap(data['document']),
-          SnapshotMetadata._(data["metadata"]["hasPendingWrites"],
-              data["metadata"]["isFromCache"]),
-          firestore,
-        ),
-        type = DocumentChangeType.values.firstWhere((DocumentChangeType type) {
-          return type.toString() == data['type'];
-        });
+  final platform.DocumentChangePlatform _delegate;
+  final Firestore _firestore;
+
+  DocumentChange._(this._delegate, this._firestore) {
+    platform.DocumentChangePlatform.verifyExtends(_delegate);
+  }
 
   /// The type of change that occurred (added, modified, or removed).
-  final DocumentChangeType type;
+  platform.DocumentChangeType get type => _delegate.type;
 
   /// The index of the changed document in the result set immediately prior to
   /// this [DocumentChange] (i.e. supposing that all prior DocumentChange objects
   /// have been applied).
   ///
   /// -1 for [DocumentChangeType.added] events.
-  final int oldIndex;
+  int get oldIndex => _delegate.oldIndex;
 
   /// The index of the changed document in the result set immediately after this
   /// DocumentChange (i.e. supposing that all prior [DocumentChange] objects
   /// and the current [DocumentChange] object have been applied).
   ///
   /// -1 for [DocumentChangeType.removed] events.
-  final int newIndex;
+  int get newIndex => _delegate.newIndex;
 
   /// The document affected by this change.
-  final DocumentSnapshot document;
+  DocumentSnapshot get document =>
+      DocumentSnapshot._(_delegate.document, _firestore);
 }
