@@ -4,26 +4,17 @@
 
 part of cloud_firestore;
 
-@visibleForTesting
-enum FieldValueType {
-  arrayUnion,
-  arrayRemove,
-  delete,
-  serverTimestamp,
-  incrementDouble,
-  incrementInteger,
-}
-
 /// Sentinel values that can be used when writing document fields with set() or
 /// update().
-class FieldValue {
-  FieldValue._(this.type, this.value);
+///
+/// This class serves as a static factory for [FieldValuePlatform] instances, but also
+/// as a facade for the [FieldValue] type, so plugin users don't need to worry about
+/// the actual internal implementation of their [FieldValue]s after they're created.
+class FieldValue extends platform.FieldValuePlatform {
+  static final platform.FieldValueFactoryPlatform _factory =
+      platform.FieldValueFactoryPlatform.instance;
 
-  @visibleForTesting
-  final FieldValueType type;
-
-  @visibleForTesting
-  final dynamic value;
+  FieldValue._(this._delegate) : super(_delegate);
 
   /// Returns a special value that tells the server to union the given elements
   /// with any array value that already exists on the server.
@@ -33,7 +24,7 @@ class FieldValue {
   /// will be overwritten with an array containing exactly the specified
   /// elements.
   static FieldValue arrayUnion(List<dynamic> elements) =>
-      FieldValue._(FieldValueType.arrayUnion, elements);
+      FieldValue._(_factory.arrayUnion(_CodecUtility.valueEncode(elements)));
 
   /// Returns a special value that tells the server to remove the given
   /// elements from any array value that already exists on the server.
@@ -42,27 +33,31 @@ class FieldValue {
   /// If the field being modified is not already an array it will be overwritten
   /// with an empty array.
   static FieldValue arrayRemove(List<dynamic> elements) =>
-      FieldValue._(FieldValueType.arrayRemove, elements);
+      FieldValue._(_factory.arrayRemove(_CodecUtility.valueEncode(elements)));
 
   /// Returns a sentinel for use with update() to mark a field for deletion.
-  static FieldValue delete() => FieldValue._(FieldValueType.delete, null);
+  static FieldValue delete() => FieldValue._(_factory.delete());
 
   /// Returns a sentinel for use with set() or update() to include a
   /// server-generated timestamp in the written data.
   static FieldValue serverTimestamp() =>
-      FieldValue._(FieldValueType.serverTimestamp, null);
+      FieldValue._(_factory.serverTimestamp());
 
   /// Returns a special value for use with set() or update() that tells the
   /// server to increment the field’s current value by the given value.
-  static FieldValue increment(num value) {
-    // It is a compile-time error for any type other than int or double to
-    // attempt to extend or implement num.
-    assert(value is int || value is double);
-    if (value is double) {
-      return FieldValue._(FieldValueType.incrementDouble, value);
-    } else if (value is int) {
-      return FieldValue._(FieldValueType.incrementInteger, value);
-    }
-    return null;
+  static FieldValue increment(num value) =>
+      FieldValue._(_factory.increment(value));
+
+  dynamic _delegate;
+
+  @override
+  String toString() => '$runtimeType($_delegate)';
+
+  @override
+  bool operator ==(Object o) {
+    return o is FieldValue && o._delegate == _delegate;
   }
+
+  @override
+  int get hashCode => _delegate.hashCode;
 }
