@@ -10,103 +10,26 @@ part of cloud_firestore;
 ///
 /// Once committed, no further operations can be performed on the [WriteBatch],
 /// nor can it be committed again.
+
 class WriteBatch {
-  WriteBatch._(this._firestore)
-      : _handle = Firestore.channel.invokeMethod<dynamic>(
-            'WriteBatch#create', <String, dynamic>{'app': _firestore.app.name});
+  final platform.WriteBatchPlatform _delegate;
 
-  final Firestore _firestore;
-  Future<dynamic> _handle;
-  final List<Future<dynamic>> _actions = <Future<dynamic>>[];
-
-  /// Indicator to whether or not this [WriteBatch] has been committed.
-  bool _committed = false;
-
-  /// Commits all of the writes in this write batch as a single atomic unit.
-  ///
-  /// Calling this method prevents any future operations from being added.
-  Future<void> commit() async {
-    if (!_committed) {
-      _committed = true;
-      await Future.wait<dynamic>(_actions);
-      await Firestore.channel.invokeMethod<void>(
-          'WriteBatch#commit', <String, dynamic>{'handle': await _handle});
-    } else {
-      throw StateError("This batch has already been committed.");
-    }
+  WriteBatch._(this._delegate) {
+    platform.WriteBatchPlatform.verifyExtends(_delegate);
   }
 
-  /// Deletes the document referred to by [document].
-  void delete(DocumentReference document) {
-    if (!_committed) {
-      _handle.then((dynamic handle) {
-        _actions.add(
-          Firestore.channel.invokeMethod<void>(
-            'WriteBatch#delete',
-            <String, dynamic>{
-              'app': _firestore.app.name,
-              'handle': handle,
-              'path': document.path,
-            },
-          ),
-        );
-      });
-    } else {
-      throw StateError(
-          "This batch has been committed and can no longer be changed.");
-    }
-  }
+  Future<void> commit() => _delegate.commit();
 
-  /// Writes to the document referred to by [document].
-  ///
-  /// If the document does not yet exist, it will be created.
-  ///
-  /// If [merge] is true, the provided data will be merged into an
-  /// existing document instead of overwriting.
+  void delete(DocumentReference document) =>
+      _delegate.delete(document._delegate);
+
   void setData(DocumentReference document, Map<String, dynamic> data,
-      {bool merge = false}) {
-    if (!_committed) {
-      _handle.then((dynamic handle) {
-        _actions.add(
-          Firestore.channel.invokeMethod<void>(
-            'WriteBatch#setData',
-            <String, dynamic>{
-              'app': _firestore.app.name,
-              'handle': handle,
-              'path': document.path,
-              'data': data,
-              'options': <String, bool>{'merge': merge},
-            },
-          ),
-        );
-      });
-    } else {
-      throw StateError(
-          "This batch has been committed and can no longer be changed.");
-    }
-  }
+          {bool merge = false}) =>
+      _delegate.setData(document._delegate,
+          _CodecUtility.replaceValueWithDelegatesInMap(data),
+          merge: merge);
 
-  /// Updates fields in the document referred to by [document].
-  ///
-  /// If the document does not exist, the operation will fail.
-  void updateData(DocumentReference document, Map<String, dynamic> data) {
-    if (!_committed) {
-      _handle.then((dynamic handle) {
-        _actions.add(
-          Firestore.channel.invokeMethod<void>(
-            'WriteBatch#updateData',
-            <String, dynamic>{
-              'app': _firestore.app.name,
-              'handle': handle,
-              'path': document.path,
-              'data': data,
-            },
-          ),
-        );
-      });
-    } else {
-      throw StateError(
-          "This batch has been committed and can no longer be changed.");
-    }
-  }
+  void updateData(DocumentReference document, Map<String, dynamic> data) =>
+      _delegate.updateData(document._delegate,
+          _CodecUtility.replaceValueWithDelegatesInMap(data));
 }
