@@ -17,7 +17,7 @@ void main() {
   // Pass all uncaught errors to Crashlytics.
   FlutterError.onError = Crashlytics.instance.recordFlutterError;
 
-  runZoned<Future<void>>(() async {
+  runZoned(() {
     runApp(MyApp());
   }, onError: Crashlytics.instance.recordError);
 }
@@ -77,6 +77,18 @@ class _MyAppState extends State<MyApp> {
                       final List<int> list = <int>[];
                       print(list[100]);
                     });
+                  }),
+              FlatButton(
+                  child: const Text('Record Error'),
+                  onPressed: () {
+                    try {
+                      throw 'error_example';
+                    } catch (e, s) {
+                      // "context" will append the word "thrown" in the
+                      // Crashlytics console.
+                      Crashlytics.instance
+                          .recordError(e, s, context: 'as an example');
+                    }
                   }),
             ],
           ),
