@@ -2,7 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-#import "FirebasePerformancePlugin+Internal.h"
+#import "FLTFirebasePerformancePlugin+Internal.h"
 
 @interface FLTHttpMetric ()
 @property FIRHTTPMetric *metric;
